Completed the api

+++ app.py
@@ -4,6 +4,7 @@ import uuid
 from werkzeug.security import generate_password_hash, check_password_hash
 import jwt
 import datetime
+from functools import wraps
 
 
 app = Flask(__name__)
@@ -24,14 +25,37 @@ class User(db.Model):
 class Todo(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     text = db.Column(db.String(50), unique=True, nullable=False)
-    complete = db.Column(db.Boolean, nullable=False)
-    user_id = db.Column(db.Integer, default=False, nullable=False)
+    completed = db.Column(db.Boolean, default=False)
+    user_id = db.Column(db.Integer, nullable=False)
 
+def check_token(f):
+    @wraps(f)
+    def decorator(*args, **kwargs):
+
+      token = None
+
+      if 'x-access-tokens' in request.headers:
+         token = request.headers['x-access-tokens']
+
+      if not token:
+         return jsonify({'message': 'a valid token is missing'}), 401
+
+      try:
+         data = jwt.decode(token, app.config['SECRET_KEY'])
+         current_user = User.query.filter_by(public_id=data['public_id']).first()
+      except:
+         return jsonify({'message': 'token is invalid'}), 401
+
+      return f(current_user, *args, **kwargs)
+    return decorator
 
 
 #Setting up the user routes that only admins who have logged in have access to
 @app.route('/user/', methods=['GET'])
-def get_all_users():
+@check_token
+def get_all_users(current_user):
+    if not current_user.is_admin:
+        return jsonify({"message": "Cannot perform that function"})
     users = User.query.all()
     output = []
     for user in users:
@@ -45,7 +69,10 @@ def get_all_users():
     return jsonify({"users": output})
 
 @app.route('/user/<user_id>', methods=["GET"])
-def get_one_user(user_id):
+@check_token
+def get_one_user(current_user, user_id):
+    if not current_user.is_admin:
+            return jsonify({"message": "Cannot perform that function"})
     user = User.query.filter_by(public_id = user_id ).first()
     if user:
         user_data = {}
@@ -59,7 +86,10 @@ def get_one_user(user_id):
         return jsonify({"user": "User not found"})
 
 @app.route('/user/', methods=['POST'])
-def create_user():
+@check_token
+def create_user(current_user):
+    if not current_user.is_admin:
+            return jsonify({"message": "Cannot perform that function"})
     data = request.get_json()
     hashed_password = generate_password_hash(data['password'], method='sha256')
     new_user_public_id = str(uuid.uuid4())
@@ -71,7 +101,10 @@ def create_user():
     return jsonify({'Message': 'New User Created'})
 
 @app.route('/user/<user_id>', methods=["PUT"])
-def promote_user(user_id):
+@check_token
+def promote_user(current_user, user_id):
+     if not current_user.is_admin:
+            return jsonify({"message": "Cannot perform that function"})
      user = User.query.filter_by(public_id = user_id ).first()
      if user:
         user.is_admin = True
@@ -82,7 +115,10 @@ def promote_user(user_id):
     
 
 @app.route('/user/<user_id>', methods=["DELETE"])
-def delete_user(user_id):
+@check_token
+def delete_user(current_user, user_id):
+    if not current_user.is_admin:
+            return jsonify({"message": "Cannot perform that function"})
     user = User.query.filter_by(public_id = user_id ).first()
     if user:
         db.session.delete(user)
@@ -119,6 +155,88 @@ def login():
 
     return make_response('User not finally verified', 401, {'WWW-Authentication': 'Basic realm="Login required"'})
 
+
+# Todo routes
+@app.route('/todo/', methods=['GET'])
+@check_token
+def get_all_todo(current_user):
+    todos=None
+    if current_user.is_admin:
+        todos = Todo.query.all()
+    else: 
+        todos = Todo.query.filter_by(user_id=current_user.public_id).all()
+        
+    if todos:
+        todoOutput = []
+        for todo in todos:
+            todo_data={}
+            todo_data['id'] = todo.id
+            todo_data['text'] = todo.text
+            todo_data['completed'] = todo.completed
+            todo_data['user_id'] = todo.user_id
+
+            todoOutput.append(todo_data)        
+
+        return jsonify({"todos": todoOutput})
+    else:
+        return jsonify({"todos": "No Todo found"})
+
+@app.route('/todo/<todo_id>', methods=['GET'])
+@check_token
+def get_one_todo(current_user, todo_id):
+    todo = Todo.query.filter_by(id= todo_id, user_id=current_user.public_id).first()
+    if todo:
+            todo_data={}
+            todo_data['text'] = todo.text
+            todo_data['completed'] = todo.completed
+            todo_data['user_id'] = todo.user_id
+
+
+            return jsonify({"todos": todo_data})
+    else:
+            return jsonify({"todos": "No Todo found"})
+
+@app.route('/todo/', methods=['POST'])
+@check_token
+def create_todo(current_user):
+    data = request.get_json()
+    new_todo_text = data['text']
+    new_todo_user_id = current_user.public_id
+    new_todo = Todo(text= new_todo_text, user_id=new_todo_user_id)
+    db.session.add(new_todo)
+    db.session.commit()
+    return jsonify({
+        "message":"New Todo Added"
+    })
+
+@app.route('/todo/<todo_id>', methods=['PUT'])
+@check_token
+def complete_all_todo(current_user, todo_id):
+    todo = Todo.query.filter_by(id= todo_id, user_id=current_user.public_id).first()
+    if todo:
+        todo.completed = True
+        db.session.commit()
+        return jsonify({
+            "message":"Todo  now completed"
+        })
+    else:
+            return jsonify({"todos": "No Todo found"})
+
+@app.route('/todo/<todo_id>', methods=['DELETE'])
+@check_token
+def delete_todo(current_user, todo_id):
+     todo = Todo.query.filter_by(id= todo_id, user_id=current_user.public_id).first()
+     if todo:
+        db.session.delete(todo)
+        db.session.commit()
+        return jsonify({"todo": "Todo Deleted"})
+     else:
+        return jsonify({"todo": "Todo not found"})
+     
+
 if __name__ == "__main__":
     db.create_all()
+    # admin_user = User(public_id = "admin" , username = "admin", password=generate_password_hash("admin", method='sha256'), is_admin=True )
+    # db.session.add(admin_user)
+    # db.session.commit()
     app.run(debug=True)
